modify btc project

+++ btc_helper/btc_helper/BTC/BtceHelper.cs
@@ -35,7 +35,7 @@ class  BtceHelper
 
     static string key;
     static HMACSHA512 hash_mark;
-    static UInt32 nonce;
+    static UInt64 nonce;
     static BtceHelper()
     {
         key = SQL.get_value("btce_key");
@@ -90,7 +90,7 @@ class  BtceHelper
         return sb.ToString();
     }
   
-    static UInt32 get_nonce()
+    static UInt64 get_nonce()
     {
         return nonce++;
     }
