+++ btc_helper/btc_helper/BTC/BtctradeHelper.cs
@@ -19,7 +19,7 @@ class BtctradeHelper
     static string api_key;
     static string secret_key;
     static HMACSHA512 hash_mark;
-    static UInt32 nonce;
+    static UInt64 nonce;
     static BtctradeHelper()
     {
         api_key = SQL.get_value("btctrade_api_key");
@@ -84,7 +84,7 @@ class BtctradeHelper
         return sb.ToString();
     }
   
-    static UInt32 get_nonce()
+    static UInt64 get_nonce()
     {
         return nonce++;
     }
