+++ btc_helper/btc_helper/BTC/HuobiHelper.cs
@@ -24,7 +24,7 @@ class HuobiHelper
     static string api_key;
     static string secret_key;
     static HMACSHA512 hash_mark;
-    static UInt32 nonce;
+    static UInt64 nonce;
     static HuobiHelper()
     {
         api_key = SQL.get_value("huobi_api_key");
@@ -89,7 +89,7 @@ class HuobiHelper
         return sb.ToString();
     }
   
-    static UInt32 get_nonce()
+    static UInt64 get_nonce()
     {
         return nonce++;
     }
